Implement version, report_issue commands
Adds version and report_issue commands.
version: Responds with the current relay extension version.
report_issue: Responds with a link a new issue page on github.
resolves #4

Also refactors the common commands code out for future reuse and
adds SNAPSHOT to the current version.
