+++ hangouts/firefox/build.gradle
@@ -1,4 +1,4 @@
-version = "0.1.0"
+version = "0.1.0-SNAPSHOT"
 ext.bowerDestination = "${buildDir}/data"
 
 task runExtension(type: Exec) {
