+++ hangouts/firefox/src/main/ts/data/commands.ts
@@ -1,0 +1,1 @@
+../../../../../../common/commands.ts
