+++ hangouts/firefox/src/main/ts/index.ts
@@ -1,7 +1,12 @@
 ///<reference path="../local-typings/page-mod.d.ts" />
+///<reference path="../local-typings/self.d.ts" />
 import pageMod = require('sdk/page-mod');
+import self = require('sdk/self');
 
 pageMod.PageMod({
     include: /.*talkgadget\.google\.com.*/,
-    contentScriptFile: ['./jquery.js', './bot.js']
+    contentScriptFile: ['./jquery.js', './commands.js', './bot.js'],
+    contentScriptOptions: {
+        version: self.version
+    }
 });
